retriever/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
numpy
pydantic
scikit-learn